requirements.txt
aiohttp                   3.8.5
aiosignal                 1.3.1
altair                    5.0.1
anyio                     3.7.1
async-timeout             4.0.3
attrs                     23.1.0
blinker                   1.6.2
boto3                     1.28.27
botocore                  1.31.27
cachetools                5.3.1
certifi                   2023.7.22
charset-normalizer        3.2.0
click                     8.1.6
colorama                  0.4.6
dataclasses-json          0.5.14
exceptiongroup            1.1.3
faiss-cpu                 1.7.4
filelock                  3.12.2
frozenlist                1.4.0
fsspec                    2023.6.0
gitdb                     4.0.10
GitPython                 3.1.32
greenlet                  2.0.2
grpcio                    1.57.0
grpcio-tools              1.57.0
h11                       0.14.0
h2                        4.1.0
hpack                     4.0.0
httpcore                  0.17.3
httpx                     0.24.1
huggingface-hub           0.14.1
hyperframe                6.0.1
idna                      3.4
importlib-metadata        6.8.0
InstructorEmbedding       1.0.1
Jinja2                    3.1.2
jmespath                  1.0.1
joblib                    1.3.2
jsonschema                4.19.0
jsonschema-specifications 2023.7.1
langchain                 0.0.184
langsmith                 0.0.22
markdown-it-py            3.0.0
MarkupSafe                2.1.3
marshmallow               3.20.1
mdurl                     0.1.2
mpmath                    1.3.0
multidict                 6.0.4
mypy-extensions           1.0.0
networkx                  3.1
nltk                      3.8.1
numexpr                   2.8.5
numpy                     1.25.2
openai                    0.27.8
openapi-schema-pydantic   1.2.4
packaging                 23.1
pandas                    2.0.3
Pillow                    9.5.0
pip                       24.0
portalocker               2.7.0
protobuf                  4.24.0
pyarrow                   12.0.1
pydantic                  1.10.12
pydeck                    0.8.0
Pygments                  2.16.1
Pympler                   1.0.1
PyPDF2                    3.0.1
python-dateutil           2.8.2
python-dotenv             1.0.0
pytz                      2023.3
pytz-deprecation-shim     0.1.0.post0
pywin32                   306
PyYAML                    6.0.1
qdrant-client             1.4.0
referencing               0.30.2
regex                     2023.8.8
requests                  2.31.0
rich                      13.5.2
rpds-py                   0.9.2
s3transfer                0.6.2
sacremoses                0.0.53
safetensors               0.3.2
scikit-learn              1.3.0
scipy                     1.11.1
sentence-transformers     2.2.2
sentencepiece             0.1.99
setuptools                49.2.1
six                       1.16.0
smmap                     5.0.0
sniffio                   1.3.0
SQLAlchemy                2.0.19
streamlit                 1.33.0
sympy                     1.12
tenacity                  8.2.3
threadpoolctl             3.2.0
tiktoken                  0.4.0
tokenizers                0.13.3
toml                      0.10.2
toolz                     0.12.0
torch                     2.0.1
torchvision               0.15.2
tornado                   6.3.3
tqdm                      4.66.1
transformers              4.31.0
typing_extensions         4.7.1
typing-inspect            0.9.0
tzdata                    2023.3
tzlocal                   4.3.1
urllib3                   1.26.16
validators                0.21.2
watchdog                  3.0.0
yarl                      1.9.2
zipp                      3.16.2